changes to make betarun/v2.8 compile for sparc
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@2732 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/argument.c
@@ -54,7 +54,7 @@ void TraceXcall(long arg1, ...)
   fprintf(output, ")\n"); fflush(output);
 }
 
-void TraceXres(long res)
+void TraceXres(unsigned int res)
 {
   fprintf(output, "returns 0x%x\n", res);
   fflush(output);
